backend/requirements.txt
# Core
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# HTTP & extraction
requests==2.31.0
readability-lxml==0.8.1
lxml==5.1.0

# AI
google-generativeai>=0.8.0

# DB & utils
aiosqlite==0.19.0
pydantic==2.6.1
pydantic-settings==2.1.0
